Refactor with logging and db connection handling
