+++ server/sql/005_add_grants.down.sql
@@ -1,0 +1,2 @@
+REVOKE GRANT OPTION FOR ALL PRIVILEGES ON TABLE addresses FROM pqvp;
+REVOKE GRANT OPTION FOR ALL PRIVILEGES ON TABLE profiles FROM pqvp;
