+++ server/sql/005_add_grants.up.sql
@@ -1,0 +1,2 @@
+GRANT ALL PRIVILEGES ON TABLE addresses TO pqvp;
+GRANT ALL PRIVILEGES ON TABLE profiles TO pqvp;
