+++ server/src/pqvp/db.go
@@ -5,6 +5,7 @@ import (
 	"io"
 
 	_ "github.com/lib/pq"
+	"github.com/paulmach/go.geojson"
 	"golang.org/x/crypto/bcrypt"
 )
 
@@ -44,6 +45,42 @@ func (pg *Postgres) DeleteUser(u User) error {
 	return nil
 }
 
+// FindRecipients handle the actual PostGIS query
+// looking for all addresses inside the alert geometry
+func (pg *Postgres) FindRecipients(geo *geojson.Geometry) ([]string, error) {
+	var phoneNumbers []string
+
+	json, err := geo.MarshalJSON()
+	if err != nil {
+		return nil, err
+	}
+
+	// Look for the phone numbers that having addresses in the alert geometry
+	// Right now we assume the GeoJSON geometry is a polygon
+	query := `
+	   SELECT DISTINCT p.phone
+	   FROM Addresses a, Profiles p
+	   WHERE ST_Intersects(a.point, ST_SetSRID(ST_GeomFromGeoJSON($1),4326))
+	   AND a.profile_id = p.id`
+
+	rows, err := pg.Query(query, json)
+	if err != nil {
+		return phoneNumbers, err
+	}
+
+	for rows.Next() {
+		var phoneNumber string
+		err = rows.Scan(&phoneNumber)
+		phoneNumbers = append(phoneNumbers, phoneNumber)
+	}
+	if err != nil {
+		return nil, err
+
+	}
+
+	return phoneNumbers, nil
+}
+
 // Close implements io.Closer
 func (pg *Postgres) Close() error {
 	return pg.DB.Close()
@@ -57,6 +94,7 @@ type Datastore interface {
 	LoginUser(User) bool
 	CreateUser(User) error
 	DeleteUser(User) error
+	FindRecipients(*geojson.Geometry) ([]string, error)
 	io.Closer
 }
